petsitter ê 기능 추가(#74)

+++ BE/src/main/java/shop/petmily/domain/reservation/dto/ReservationPossiblePetsitterReseponseDto.java
@@ -34,5 +34,5 @@ public class ReservationPossiblePetsitterReseponseDto {
 
     private Time possibleTimeEnd;
 
-    private int star;
+    private double star;
 }
